requirements.txt
tweepy
pandas
python-binance
prophet
newsapi-python
requests
streamlit
ollama
python-dotenv
psycopg2-binary
plyer
newspaper3k
lxml_html_clean
beautifulsoup4
fuzzywuzzy
python-Levenshtein
websocket-client
aiohttp
fastapi
uvicorn
websockets
pydantic
yfinance
pmdarima
joblib